Moved TagInput to its own directory.
